pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Rob Caduyac",email = "rmcaduyac1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain-community (>=0.3.19,<0.4.0)",
    "pypdf (>=5.3.1,<6.0.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langchain-pinecone (>=0.2.3,<0.3.0)",
    "langgraph (>=0.3.9,<0.4.0)",
    "langgraph-checkpoint-postgres (>=2.0.17,<3.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "langsmith (>=0.3.18,<0.4.0)",
    "psycopg[binary,pool] (>=3.2.6,<4.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "chainlit (>=2.4.400,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
